fixed up serial + UART + I2C chapters, cleaned up diagrams
The description of serial ports and serial communication was
pretty confusing: now it's much cleaner. The new chapter
name `Serial Port` should help, as it unambiguously
describes the thing we will actually be working with.

The I2C chapter had some slightly weird stuff going on that
was fixed up a bit.

The handling of figures was cleaned up. A Wikimedia Commons
diagram that was incorrectly copied instead of linked was
fixed. Figure background was added and slightly darkened
with additional CSS to work reasonably with dark-mode
reading.

Thanks much to Github @iampi31 for contributions and
feedback that became part of this PR.

closes #78, closes #79, closes #81